@@ -5,7 +5,7 @@
   "scripts": {
     "ng": "ng",
     "build": "ng build",
-    "build:jit": "ng build --localize=false --ts-config=tsconfig-jit.app.json --main=src/main-jit.ts",
+    "build:jit": "ng build --localize=false --ts-config=tsconfig-jit.app.json --browser=src/main-jit.ts",
     "start": "ng serve",
     "start:he": "ng serve -c=he",
     "start:uk": "ng serve -c=uk",
@@ -19,24 +19,24 @@
     "extract-i18n-merge:xlf2": "ng run app:extract-i18n-merge:xlf2"
   },
   "dependencies": {
-    "@angular/animations": "17.0.3",
-    "@angular/common": "17.0.3",
-    "@angular/compiler": "17.0.3",
-    "@angular/core": "17.0.3",
-    "@angular/forms": "17.0.3",
-    "@angular/platform-browser": "17.0.3",
-    "@angular/platform-browser-dynamic": "17.0.3",
-    "@angular/router": "17.0.3",
+    "@angular/animations": "17.2.4",
+    "@angular/common": "17.2.4",
+    "@angular/compiler": "17.2.4",
+    "@angular/core": "17.2.4",
+    "@angular/forms": "17.2.4",
+    "@angular/platform-browser": "17.2.4",
+    "@angular/platform-browser-dynamic": "17.2.4",
+    "@angular/router": "17.2.4",
     "rxjs": "7.8.0",
     "tslib": "2.3.0",
     "zone.js": "0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.0.1",
-    "@angular/cli": "17.0.1",
-    "@angular/compiler-cli": "17.0.3",
-    "@angular/localize": "17.0.3",
-    "ng-extract-i18n-merge": "2.9.0",
+    "@angular-devkit/build-angular": "17.2.3",
+    "@angular/cli": "17.2.3",
+    "@angular/compiler-cli": "17.2.4",
+    "@angular/localize": "17.2.4",
+    "ng-extract-i18n-merge": "2.11.0",
     "typescript": "5.2.2"
   }
 }
