@@ -2,7 +2,7 @@ import { getLocaleDirection } from '@angular/common';
 
 import { Locale } from '../locale';
 
-export const setDocumentI18nAttributes = (locale: Locale): void => {
+export const setDocumentI18nAttributes = (locale: Locale) => {
   /**
    * Can be replaced with native `Intl.Locale.prototype.textInfo` when Firefox will support it.
    * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/Locale/textInfo
